Backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
bleach==6.1.0
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
django-ipware==7.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-recaptcha==4.0.2
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pillow==10.4.0
PyJWT==2.9.0
python-ipware==3.0.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
webencodings==0.5.1
